backend/requirements.txt
Flask>=2.0
requests>=2.28
flask-cors>=3.0
python-dotenv>=1.0
pytest>=7.0
flask-limiter>=3.0
gunicorn==21.2.0
